ItemMapping: Use the base variant ID for variant mappings
This change allows entries to be more readable, like from the base
tradeable item to the most common charged or modified variant, rather
than requiring the base variant be given as the "untradeable item" to be
mapped.

eg. Instead of requiring the form

```
ITEM_RING_OF_SUFFERING(RING_OF_SUFFERING, true, 1L, RING_OF_SUFFERING),
```

this entry can map to `RING_OF_SUFFERING_I`, or any other possible
variant of the ring of suffering. At the time of writing, this affects
most Nightmare Zone imbued rings, the Black mask, the Ring of wealth,
and the Bow of faerdhinen.