menu swapper: Fix duplicate Reset option in bank
Because the bank tags plugin adds an "Edit-tags" menu entry which has a
null widget, this causes the highestOp calculation to be incorrect for
all entries prior to it, as the loop effectively `break`s when some menu
entry with a widget not matching that of the current entry's widget is
encountered. This commit fixes this behavior by moving the widget check
inside the loop body and `continue`ing on a non-match rather than
`break`ing.