idlenotifier: handle non-standard attack options during npc transformations
NPCs which begin combat with a menu option that is not "Attack" are missed during combat detection. This PR uses NpcChanged to detect when an NPC that we are interacting with transforms and gains an Attack option while still being interacted with.

I noticed this at Kraken since combat begins with "Disturb" as a Whirlpool before the NPC transforms into the Kraken, and verified that my change resolved the issue.