timers: Track Arceuus spell durations using varbits
The resurrect thrall cooldown duration was updated to 17 ticks from 12
ticks in line with the enhanced client buff bar[1], because this var
changes 5 ticks before the game message is sent.

Ward of Arceuus must still be triggered by a chat message because the
var controlling it is not transmitted to the Java client.

[1]: https://github.com/Joshua-F/cs2-scripts/blob/1202119848818ef90e38d488f64504ecf491d845/scripts/%5Bproc%2Cbuff_bar_get_value%5D.cs2#L180-L183