api: add rl object model supplier + animation controller
Animations applied to models did not account for scale, leading to problems when applying them to RL objects which needed to be shrunk/enlarged. For example, the Nightmare pet, which uses the same model as the regular Nightmare, but with scale(30, 30, 30).

This PR gives the ability for plugin developers to supply their own model on each frame, and exposes animation methods to the api as well as an AnimationController to simplify usage.