api: add rl object post-animation transform
Animations applied to models did not account for scale, leading to problems when applying them to RL objects which needed to be shrunk/enlarged. For example, the Nightmare pet, which uses the same model as the regular Nightmare, but with scale(30, 30, 30).

This PR gives the ability for plugin developers to modify the model after it has been passed through the animation transform, allowing arbitrary further modifications.