config manager: reset lost profile active flag
In the event the profiles were simply lost on disk, this would
otherwise cause the client to switch to a new default profile
as no profile would be active.

It is not safe to reset sync/rev due to if the profile was actually
deleted by another client, the remote profile would be deleted too,
causing it to re-sync the partially-patched remote profile.

(cherry picked from commit c0b9105d6b891a4bf06767b97d0f870588085927)