Update `Show this to Hazelmere`
Moved the worldPoint to 0 so it's on the ground and can actually point to the stairs, when it's set to 1 it can never trigger due to the NPC being in the same area.

https://github.com/Felanbird/runelite/assets/41973452/b29bfd9c-83f0-4574-a946-66e8af9cea7b
https://github.com/Felanbird/runelite/assets/41973452/e604ef33-933e-4ec2-bc64-9d35880e32c1

note: I'm aware the images in this commit are not for this Hazelmere step, but the logic applies to this step as well.