api: worldpoint: use client base x/y instead of scene
Map loads under vanilla processing order looks roughly like:

tick 0:
- map load packet. client changes to LOADING state.
- shift locations of players, npcs, etc.
- assign new base x/y
- any other packets sent within the same tick
  - player update
  - stat change
  - etc

tick 1:
- map load code runs. this might complete the map load and move the
  client to LOGGED_IN state.
- LOGGED_IN tick runs as normal

Notably there are several packets which get processed during LOADING in
tick 0 after the scene has shifted, but before the new scene loads.
These packets assume the new base x/y is applied and send coordinates
relative to it. This causes client base x/y to be different from scene
base x/y during this time.

The behavior of the worldpoint methods which accept a Client was
inadvertently changed when adding the overloads for Scene, due to
changing from client.getBaseXY() to client.getScene().getBaseXY() in
37580e5c72221b067e3c02d3d0131e0e8d5d4c75 and
09b78ad626c8ca4034fab4706e31311045055ac2.

This reverts to the previous behavior.

This was breaking the agility plugin counting laps at the agility
pyramid. The plugin checks the player location during the stat change
event to determine if a lap has been completed. This stat change event
happens when the packet is processed, which is in between scene shift
and scene load, causing it to use the previous scene's base instead,
leading to an invalid player location to be computed.