overlay renderer: don't consume mouse drag events when moving overlays
This was added with resizable overlays in
d676542dc2f741ba1d821e73d3e7cf9f67df9626, but I think it was just copied
from the mouse click handler. Consuming the mouse click is enough to
prevent the vanilla client from rotating the camera due to mouse drag.

Consuming the drag event causes real mouse position to desync from where
the client thinks it is, which gives odd behaviors like it causes the
minimenu to instantly close if opened, since the client thinks the mouse
is elsewhere.