config: fix duplicating rsprofile
Even though the web service v2 to v3 migration names the $rsprofile
correctly, profiles created since the addition of the v3 api were never
named.

Commonly, this caused a duplication of profile -1. During profile sync,
the client would apply the remote profile name "" to $rsprofile, and
then later recreate $rsprofile due to no profile with the name existing.

The config manager rsProfile was then this newly created $rsProfile, so
the disk storage of the profile has been in the correct location.

The "" profile then shows up in the profile panel, and deleting it would
then delete both it and the $rsprofile profile - causing the config
manager to recreate it due to having lost its active profile.

This detects and removes duplicate profiles, and corrects the name of
the rsprofile profile.

I've also included the profile name in the config patch, so the name of
internal profiles are set correctly. Newly synced profiles also uses
this now too. Existing profiles still uses the rename api call to avoid
clients fighting with each other over the profile name if it is changed
on one of them but not the other.

This also fixes trying to sync newly created profiles with no settings, which
previously 400d due to posting a null patch. This caused the profile to
not really be synced and be considered "lost" at the next profile sync.

(cherry picked from commit d64092153f035dbbae0012d871287b6edd00ccc6)