gpu: remove sin/cos table from UBO
This was added in 9c791f5b09d0defe7c357fb986ad5600de9e3dfd due to
inconsistencies in the computed sin/cos values causing graphical issues
on some cards. However the sin/cos table is no longer used for the
projection, and instead only for computing face distance for face
sorting, so I think it is safe to remove.