default world: Remove world change on profile change
In the past, config profiles could only be changed manually via the
config panel, so it made sense for the world at the login screen to
change according to the new profile's configs. As of the previous
commit, it's not always the case that profile changes are triggered
directly by user--they can now occur as the result of world hops, or
even changing the selected world at the login screen. Because of this,
it is possible for the default world plugin to switch worlds in response
to the world-change-triggered-profile change, which could in turn
trigger further profile changes. To avoid this possible looping behavior,
the plugin no longer responds to config profile changes.