bank tags: refactor ui to use layers
This rewrites most of the tab interface code to use a layer for both the
tab list and also the tag tab tab. The scrolling has been reimplemented
to use native scrolling.

Also modify the bank searching script hook to check items even when the
bank is not in searching mode, which removes some complexity with having
to have the bank in a pseudo-search mode but still allow other searches
to be started.

Replace the import/export notifications with chat messages, which are
always actually delivered and aren't annoying.