clientui: correctly apply suction when expanding via min size change
when expanding the sidebar while 'keep game size' was set, and the game
width was less than the min size of the window after expanding,
rightSideSuction would not be set. This is because setting the frame's
min size causes the frame to immediately resize if it is too small, so
ContainableFrame would not see a simultaneous expand + shift like it
expects.