clientui: remember last restored bounds when starting maximized
Previously if you maximized the client, closed it, opened it, then
restored it, it would take up the whole screen because it remembered its
size from when it was maximized.

Instead save the restored position, so when the client restores after a
restart it will have a sane size & position.