Add functionality to determine map point from WorldPoint
Currently the World Map maps are not entirely accurate as to where things are, which means the dev tool for the map will display the wrong coordinates.

This adds some base functionality to handle mapping from real WorldPoints to World Map WorldPoints, with consideration for the current area viewed in the map.