Add third colour to Chat Timestamps for Split PM messages
At this time Chat Timestamps uses the chatbox's current opacity to determine what colour to use when displaying timestamps. Because of this, the messages that appear in the Split Private Message box, which appears above the chatbox, are coloured according to the setting of the chatbox. Since RuneScape offers the ability to modify the colour of the split PM text, separate from the text color of the opaque or transparent chatbox, the plugin should be updated to allow split PM messages to have a distinct timestamp colour as well. This change adds a third colour to the plugin and correctly applies it to messages from the split PM box.