mes: custom swaps for dizana's quiver ammo slot
Dizana's quiver adds a new ammo slot to the equipment tab, but:
* this slot is not actually part of the equipment inventory container, and
* probably because of that, the widgets are implemented differently as a couple nested static widgets

As a result, the existing equipment tab custom tab functionality does not recognize it as a valid swap target. However, it can be handled as a normal ui swap by just excluding it from the inventory interface id preconditions.