Fix #15761:  Plugin name highlighting gets stuck when searching
The plugin tool tip text that appears when highlighting a plugin in the
configuration menu would remain stuck if you used the search bar
and didn't move the mouse. Created a new method "hideTooltip"
that sets the variable "enabled" of the class "ToolTipManager"
to false and then back to true. It's a solution a bit "hacky" but it
was the simplest and cleanest fix that i could come up with.