Clarify NPC names config description
From personal experience, I found it confusing what I had to input in this configuration to have the plugin work as expected. I found it in the wiki, and decided to copy the text there to the description so it's easier to find without having to open a wiki page.