skill calculator: Use real levels for plank actions
Now that the various types of planks are not all listed at the top of
the construction skill actions, it doesn't make sense to claim they can
all be used at level 1. This commit amends them to assume the lowest
level of player-owned house furniture they can be used to construct.