ui: overlay: fix rendering textcomponent with embedded col tag
Rendering a string like ab<col=ff0000>cd would throw due to the split()
returning "ab" as the first token, and being unable to parse a col tag
from it.

This also fixes a bug where the returned width included the col tags,
and, removes the colorWithAlpha() call which I think should not be
necessary.