mlm: update for sack size changes
Sack size changed from 81/162 to 108/189, but can no longer be over
deposited. Use the vanilla sack size overlay because there is no longer
a whole number of remaining deposits due to the over-deposit behavior
change.

Additionally turn the vanilla overlay text red if the number of paydirt
in the inventory is at or exceeds the remaining sack capacity.