2005/2006 reworked
Goal was to get 2005/2006 to be as 1:1 to the originals as possible.
Changed shared directory for 2005/2006 to be separate.
Adjusted fixed mode highlight stones to be 1:1.
Adjusted gameframe to be 1:1, with exception of the chat.
Added icons with correct shading for the time period for 2006.
Readjusted icons for both
Added resizable for both