bank tags: fix placeholders duping items with objvars
Withdrawing an item with objvars, like ring of shadows, when you have
multiple of them, would result in both a regular ring of shadows and a
placeholder ring of shadows existing in the bank at once. The regular
item would be layouted like normal, but the placeholder was considered
to be an item in the tab not in the layout, and so was added to the
layout.

Fix by removing the placeholder id too when layouting the normal id.

Also fix drawing multiple placeholders in a layout as they will have
qty>1. Instead, assume anything with a placeholder template id is a
placeholder.