account: correctly parent JOptionPane
Without a parent component, the JOptionPane may render underneath the client window if alwaysOnTop is set. This can make the JOptionPane uninteractable while also stealing focus from the frame.

Since the Account plugin doesn't have any owned components, we don't have other options for the parent. I've exposed the ContainableFrame instance from ClientUI, but if we don't want to do that then we'll need another way to provide a parent in non-Swing contexts.

I've also fixed one incorrect use in ProfilePanel, which has its own context available.

There is another instances of null-parenting in LinkBrowser, but since it is in a static context *and* is an extensively used API option, it needs a bit more consideration.

Closes #18933